packed-backend: fsck should warn when "packed-refs" file is empty
During fsck, an empty "packed-refs" gives an error; this is unwarranted.
The runtime code paths would accept an empty "packed-refs" file, such as
"create_snapshot" would simply return the "snapshot" without checking
the content of "packed-refs".

And we should also skip checking the content of "packed-refs" when it is
empty during fsck. However, we should think about whether we need to
report something to the users in this case.

After 694b7a1999 (repack_without_ref(): write peeled refs in the
rewritten file, 2013-04-22), we would always write a header into the
"packed-refs" file where we would never create empty file since then.
Because we only create empty "packed-refs" in the very early versions,
we may tighten this rule in the future. In order to notify the users
about this, we should at least report an warning to the users.

But we need to consider the fsck message type carefully, it is not
appropriate that we use "FSCK_ERROR". This is because we would
definitely break the compatibility. Let's create a "FSCK_INFO" message
id EMPTY_PACKED_REFS_FILE" to indicate that "packed-refs" is empty.

Signed-off-by: shejialuo <shejialuo@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>